simplify solution by using `tuple_combinations(...)`
This is essentially the same thing we were doing before with
`cartesian_product(...)` except that we previously needed to do a
bunch of manual bookkeeping so that we didn't compare a number
against itself.  This involved getting enumerated iterators and
filtering out combinations based on indices.

None of that is necessary with `tuple_combinations(...)` because
that behavior is built into the iterator itself.